+++ app/(site)/donate/page.tsx
@@ -1,65 +1,64 @@
-// ExploreSol/app/donate/page.tsx
-"use client"
-import React, { useState } from 'react';
+// /app/(site)/Donate/page.tsx
+import Donate from "@/components/Donate";
+import { Metadata } from "next";
+import Image from "next/image";
+import Link from "next/link";
+
+export const metadata: Metadata = {
+  title: "Donate - Explore Solana",
+  description: "Kindly support Explore Solana platform",
+  // other metadata
+};
 
 const DonatePage = () => {
-  const [walletAddress] = useState('63t6dZ78VFW1yX7uFJV678qFbZweDpUDenjLYEsq2J7q');
-  
-  const copyToClipboard = async () => {
-    if (navigator.clipboard && window.isSecureContext) {
-      // Try to use the Clipboard API
-      try {
-        await navigator.clipboard.writeText(walletAddress);
-        alert('Wallet address copied to clipboard!');
-      } catch (err) {
-        console.error('Copy failed:', err);
-        fallbackCopyTextToClipboard(walletAddress);
-      }
-    } else {
-      // Fallback if Clipboard API is not available
-      fallbackCopyTextToClipboard(walletAddress);
-    }
-  };
-
-  const fallbackCopyTextToClipboard = (text: string) => {
-    // Create a text area element dynamically
-    const textArea = document.createElement('textarea');
-    textArea.value = text;
-    document.body.appendChild(textArea);
-    textArea.focus();
-    textArea.select();
-
-    try {
-      const successful = document.execCommand('copy');
-      const msg = successful ? 'successful' : 'unsuccessful';
-      console.log('Fallback: Copying text command was ' + msg);
-      alert('Wallet address copied to clipboard!');
-    } catch (err) {
-      console.error('Fallback: Oops, unable to copy', err);
-      alert('Failed to copy wallet address. Please try manually.');
-    }
-
-    document.body.removeChild(textArea);
-  };
-
   return (
-    <div className="container mx-auto px-4 text-center">
-      <h1 className="text-2xl font-bold text-center my-6 text-purple-400">
-        Kindly Donate to Keep ExploreSol Project Going
-      </h1>
-      <p className="text-md mb-4">
-        You can support by donating SOL and SOL tokens to the following wallet address - THANKS!:
-      </p>
-      <p className="input-address" style={{ userSelect: 'text' }}>
-        {walletAddress}
-      </p>
-      <button onClick={copyToClipboard} className="btn mt-2">
-        Copy Wallet Address
-      </button>
-    </div>
+    <>
+    <section className="flex-grow pb-12.5 pt-32.5 lg:pb-25 lg:pt-45 xl:pb-30 xl:pt-50">
+      <div className="relative z-1 mx-auto max-w-c-1016 px-7.5 pb-7.5 pt-10 lg:px-15 lg:pt-15 xl:px-20 xl:pt-20">
+        <div className="absolute left-0 top-0 -z-1 h-2/3 w-full rounded-lg bg-gradient-to-t from-transparent to-[#dee7ff47] dark:bg-gradient-to-t dark:to-[#252A42]">
+          <div className="absolute bottom-17.5 left-0 -z-1 h-1/3 w-full">
+            <Image
+              src="/images/shape/shape-dotted-light.svg"
+              alt="Dotted"
+              className="dark:hidden"
+              fill
+            />
+            <Image
+              src="/images/shape/shape-dotted-dark.svg"
+              alt="Dotted"
+              className="hidden dark:block"
+              fill
+            />
+          </div>
+
+          <div className="flex flex-col items-center justify-center">
+            <Link
+              href="/"
+              className="text-foreground bg-btn-background hover:bg-btn-background-hover group absolute left-8 top-8 flex items-center rounded-md px-4 py-2 text-sm no-underline"
+            >
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                width="24"
+                height="24"
+                viewBox="0 0 24 24"
+                fill="none"
+                stroke="currentColor"
+                strokeWidth="2"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                className="mr-2 h-4 w-4 transition-transform group-hover:-translate-x-1"
+              >
+                <polyline points="15 18 9 12 15 6" />
+              </svg>{" "}
+              Back
+            </Link>
+          </div>
+        </div>
+            <Donate />
+      </div>
+    </section>
+    </>
   );
 };
 
 export default DonatePage;
-
-
